fleshing out parser implementation

+++ projects/06/assembler/parser/tokens.hh
@@ -2,13 +2,13 @@
 
 // ASCII TABLE REF: cs.cmu.edu/~pattis/15-1XX/common/handouts/ascii.html
 
-uint8_t EQUAL = 61;         // '='
-uint8_t PLUS = 43;          // '+'
-uint8_t MINUS = 45;         // '-'
-uint8_t AT_SIGN = 64;       // '@'
-uint8_t SEMICOLON = 59;     // ';'
-uint8_t LEFT_BRACKET = 40;  // '('
-uint8_t RIGHT_BRACKET = 41; // ')'
-uint8_t A_INSTRUCTION = 0;  // in Hack Architecture @XXX
-uint8_t L_INSTRUCTION = 1;  // in Hack Architecture (XXX)
-uint8_t C_INSTRUCTION = 2;  // in Hack Architecture dest=comp;jump
+int EQUAL = 61;         // '='
+int PLUS = 43;          // '+'
+int MINUS = 45;         // '-'
+int AT_SIGN = 64;       // '@'
+int SEMICOLON = 59;     // ';'
+int LEFT_BRACKET = 40;  // '('
+int RIGHT_BRACKET = 41; // ')'
+int A_INSTRUCTION = 0;  // in Hack Architecture @XXX
+int L_INSTRUCTION = 1;  // in Hack Architecture (XXX)
+int C_INSTRUCTION = 2;  // in Hack Architecture 'dest=comp;jump'
